+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## 0.11.1
+## 0.12.0
 
+* Adding the `sampler` options to `clustering/k-means`.
+* Fixing a bug with `clustering/k-means` where k would be superior to the number of vectors.
 * Fixing a bug with `clustering/k-means` `initialCentroids` options.
+* Fixing a bug with `clustering/k-means` where a vector could end up in several clusters.
 
 ## 0.11.0